Created the project structure, and structure of the main classes
To the entities, I added the attributes I thought would be neccessary, but those are highly subject to change.

Also in the entities, they have the entity tag, but commented out. This is because they are entities, but giving them the tag neccessitates properly relating them to one another, which we don't need to have yet, so I left it commented.