Ingredients now have their own page
The ingredients appear with their photo, name and description and three sections that also appear in the nav bar

The first is the ingredients with which it combines.

The second is recipes that contain it.

Both with a button to configure recipes or ingredients

The last section is the actions section where there is the button to edit and delete ingredients that are not yet functional.

Also adapted the recipe footer to our page.

Both Web and Rest controllers can view the details of an ingredient with its ID. Currently there may be circular dependencies.