Implemented Recipe Creation
The same template should be copied for updating recipes.

Showing the recipe back on Recipes still needs some work. mainly, since the placeholder images are in /img/, the template uses /img/{{recipe.img}}, but that doesn't work when the image is a full url.

Also did some other work on other issues.