Developed Recipe Rest Controller methods
Outside of modification, which I haven't tested, everything seems to work as expected.

Creating and editing should be looked at in detail. I am unsure as to how a recipe with ingredients would be created, should you put the whole ingredient in the body? What if it doesn't exist? Maybe pivoting to using ids would be the smarter choice.

Again, creation and editing should be looked at in further detail.