Implemented DTOs for each of the entities
This DTO are copies of their respective class, but where the original class had a reference to an entity, the DTO has a Long representing its foreign key.

Also, a DTO can be created from its entity, by doing, for example new RecipeDTO(recipe). It will be created to match the provided entity. They can also be created through an empty constructor and getters/setters.

Likely, a new constructor has been added to each entity, to create it from a DTO. HOWEVER, the attributes which are references to other entities are NOT copied, as this would neccesitate having access to the corresponding service, and apparently giving a model access to a service is a bad practice. So when using this constructor, one should keep in mind that they still need to call the setters for any attributes that reference other entities.

All instances of entities being sent to or received from frontend have been changed to send/receive a DTO. I have applied some simple logic to change from the DTO to the entity (when its received) and from the entity to the DTO (when it is sent). This should work. However, further testing is neccessary, as I'm not very comfortable testing through postman and I haven't done many tests.

The DTOs have a function toDTO, which takes a collection of its corresponding entity, and returns a collection of its DTO.

Likewise, a variant of getElementById was created in each service, to recive a collection of ids and return a collection formed by the entities matching those ids.

For consistency's sake, the field img of class Recipe was changed to image. This neccessitated changing multiple views that used recipes to match the new attribute name.

There are a series of requests for Recipe instances done from frontend. These have not been changed to DTOs, as it seems that it will be slightly more complex, since it will need changes on how the frontend prepares the data to be sent. As such, these changes are left to be made in successive commits.