Solved some issues with Rest and DTOs
Mainly issues having to do with null values existing where they shouldn't.

Also, added some validation to check that the recipes to add/modify are valid.

Did some further testing on recipe CRUD through the API, and all seems to work correctly.

Found an issue when creating a recipe with no chef, where the recipe can't be displayed in the web. I found the issue and will solve it in consequent commits.