Fixed recipe creation and editing on Web
Previously, the recipe to be set was received as a String and manually mapped to a Recipe object. It was done this way because I had no understanding of the way Jackson works. Now, having investigated it and understood it, I was able to change it.

Now it receives a RecipeDTO, and creates the Recipe from that. It has been taken into consideration that the RecipeDTO received here could have ingredients and chef, as opposed to in REST, where they won't have those.

Since how the data is received and, to some extent, treated, has changed, it was also neccesary to tweak the js which generates that data so that it produces it with the correct format.