Integrated recipes with the database
Recipes are stored in the database upon creation, and changes to the stored recipes will be reflected in the database.

When a recipe is created or edited, its corresponding chef will be edited to have said recipe in its list of recipes.

Small breakdown of the changes to each file:
Chef and ChefService have been modified to add necesary/convenient functionality.

RecipeController and RecipeRestController weremodified to use Optional, since the previous implementation did not use it. Also, they were modified where needed to fix oversights or errors.

CreateRecipe was edited to add back the functionality to check that the recipe has a chef and at least one ingredient.

RecipeUpdateForm was modified to remove the option creating a recipe with no chef.