Added consequences for deleting an entity
Ingredients can not be deleted unless it is used in no recipes.

Recipes can ALWAYS be deleted.

Chefs can always be deleted. If a chef had recipes, these will be given to a custom default chef. Ideally we could give it a special id like -1 or 0.

I have done some basic teting, and this implementation seems to work properly. However, additional testing would be ideal.

Also, I would like to take a closer look to when exactly each change is saved to the database, as well as which side of each relationship is the owner and such things.